chatbot_backend/pyproject.toml
[project]
name = "chatbot_backend"
version = "0.1.0"
description = "chatbot_backend using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.177.0,<1.0.0",
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "pydantic",
    "pyyaml",
    "requests",
    "langchain-openai", # or another provider
    "qdrant-client"     # or another vector DB client
]

[project.scripts]
chatbot_backend = "chatbot_backend.main:run"
run_crew = "chatbot_backend.main:run"
train = "chatbot_backend.main:train"
replay = "chatbot_backend.main:replay"
test = "chatbot_backend.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"